backend/requirements.txt
Django==1.10.5
fabric3
django-cors-headers==2.1.0
django-filter==1.0.1
django-rest-swagger==2.1.1
django-tz-detect==0.2.8
djangorestframework==3.5.4
ecdsa==0.13
paramiko==2.1.1
psycopg2==2.6.2
pycrypto==2.6.1
pytz==2016.10
raven==6.0.0
six==1.10.0
uWSGI==2.0.14
